+++ NEWS.md
@@ -6,6 +6,16 @@
    This means that as soon as it has been called, environment variables 
    such as `NSLOTS` will no longer be queried.
 
+ * Starting with R 4.5.0 (currently R-devel), one can use
+   `parallel::makeCluster(n, type = parallelly::RPSOCK)` as an
+   alternative to `parallelly::makeClusterPSOCK(n)`.  Similarly, `type
+   = parallelly::RMPI` creates a cluster using
+   `parallelly::makeClusterMPI()`, and `type = parallelly::SEQ`
+   creates a cluster using `parallelly::makeClusterSequential()`.
+   This was first introduced in **parallelly** 1.38.0, but here we rename
+   `PSOCK` to `RPSOCK` and `MPI` to `RMPI` to minimize the risk for
+   mistaking them from the built-in types in the **parallel** package.
+
 ## Documentation
 
  * Add more help on the R option `parallelly.maxWorkers.localhost`